+++ packages/pdfx/lib/src/renderer/io/platform_method_channel.dart
@@ -154,7 +154,7 @@ class PdfPageMethodChannel extends PdfPage {
     String? backgroundColor,
     Rect? cropRect,
     int quality = 100,
-    bool printQuality = false,
+    bool forPrint = false,
     @visibleForTesting bool removeTempFile = true,
   }) =>
       _lock.synchronized<PdfPageImage?>(() async {
@@ -173,7 +173,7 @@ class PdfPageMethodChannel extends PdfPage {
           backgroundColor: backgroundColor,
           crop: cropRect,
           quality: quality,
-          printQuality: printQuality,
+          forPrint: forPrint,
           removeTempFile: removeTempFile,
         );
       });
@@ -240,7 +240,7 @@ class PdfPageImageMethodChannel extends PdfPageImage {
     required String? backgroundColor,
     required Rect? crop,
     required int quality,
-    required bool printQuality,
+    required bool forPrint,
     required bool removeTempFile,
   }) async {
     if (format == PdfPageImageFormat.webp &&
@@ -268,7 +268,7 @@ class PdfPageImageMethodChannel extends PdfPageImage {
       'crop_height': crop?.height.toInt(),
       'crop_width': crop?.width.toInt(),
       'quality': quality,
-      'printQuality': printQuality,
+      'forPrint': forPrint,
     });
 
     if (obj is! Map<dynamic, dynamic>) {