+++ packages/pdfx/lib/src/renderer/io/platform_pigeon.dart
@@ -146,7 +146,7 @@ class PdfPagePigeon extends PdfPage {
     String? backgroundColor,
     Rect? cropRect,
     int quality = 100,
-    bool printQuality = false,
+    bool forPrint = false,
     @visibleForTesting bool removeTempFile = true,
   }) =>
       _lock.synchronized<PdfPageImage?>(() async {
@@ -165,7 +165,7 @@ class PdfPagePigeon extends PdfPage {
           backgroundColor: backgroundColor,
           crop: cropRect,
           quality: quality,
-          printQuality: printQuality,
+          forPrint: forPrint,
           removeTempFile: removeTempFile,
         );
       });
@@ -237,7 +237,7 @@ class PdfPageImagePigeon extends PdfPageImage {
     required String? backgroundColor,
     required Rect? crop,
     required int quality,
-    required bool printQuality,
+    required bool forPrint,
     required bool removeTempFile,
   }) async {
     if (format == PdfPageImageFormat.webp &&
@@ -265,7 +265,7 @@ class PdfPageImagePigeon extends PdfPageImage {
       ..cropWidth = crop?.width.toInt()
       ..cropHeight = crop?.height.toInt()
       ..quality = quality
-      ..printQuality = printQuality);
+      ..forPrint = forPrint);
 
     final retWidth = result.width, retHeight = result.height;
     late final Uint8List pixels;